requirements.txt
# CannaBot Dependencies
# Core Discord bot functionality
discord.py>=2.6.0

# Database connectivity
asyncpg>=0.28.0
aiosqlite>=0.19.0
psycopg2-binary>=2.9.0

# Data processing (for Leafly strain database)
pandas>=2.0.0
numpy>=1.24.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP client for API calls
aiohttp>=3.8.0

# Date/time utilities
pytz>=2023.3
python-dateutil>=2.8.0

# Development dependencies (optional)
# Uncomment for development:
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
